Refactor IdentityCreation class. Revise several closures to remove strong self references.
